refactor: modularize codebase
- Move tools to separate files in src/tools/
- Create services directory for NWS API client
- Simplify main.ts to focus on server setup
- Fix TypeScript imports and exports
- Update build process to work with new structure
