+++ src/interfaces/alertFeature.ts
@@ -1,0 +1,9 @@
+export default interface AlertFeature {
+    properties: {
+      event?: string;
+      areaDesc?: string;
+      severity?: string;
+      status?: string;
+      headline?: string;
+    };
+  }
