+++ src/interfaces/alertsResponse.ts
@@ -1,0 +1,5 @@
+import AlertFeature from "./alertFeature.js";
+
+export default interface AlertsResponse {
+    features: AlertFeature[];
+}
