+++ src/interfaces/forecastPeriod.ts
@@ -1,0 +1,8 @@
+export default interface ForecastPeriod {
+  name?: string;
+  temperature?: number;
+  temperatureUnit?: string;
+  windSpeed?: string;
+  windDirection?: string;
+  shortForecast?: string;
+}
