+++ src/interfaces/forecastResponse.ts
@@ -1,0 +1,6 @@
+import ForecastPeriod from "./forecastPeriod.js";
+export default interface ForecastResponse {
+  properties: {
+    periods: ForecastPeriod[];
+  };
+}
