+++ src/interfaces/pointsResponse.ts
@@ -1,0 +1,5 @@
+export default interface PointsResponse {
+    properties: {
+      forecast?: string;
+    };
+}
