+++ src/tools/getAlerts.ts
@@ -1,0 +1,74 @@
+import { z } from "zod";
+import { makeNWSRequest } from "../services/nwsService.js";
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+import type AlertsResponse from "../interfaces/alertsResponse.js";
+import type AlertFeature from "../interfaces/alertFeature.js";
+
+/**
+ * Format alert data into a readable string
+ */
+function formatAlert(feature: AlertFeature): string {
+  const props = feature.properties;
+  return [
+    `Event: ${props.event || "Unknown"}`,
+    `Area: ${props.areaDesc || "Unknown"}`,
+    `Severity: ${props.severity || "Unknown"}`,
+    `Status: ${props.status || "Unknown"}`,
+    `Headline: ${props.headline || "No headline"}`,
+    "---",
+  ].join("\n");
+}
+
+/**
+ * Registers the get-alerts tool with the MCP server
+ * @param server The MCP server instance to register the tool with
+ */
+export function registerAlertsTool(server: McpServer): void {
+  server.tool(
+    "get-alerts",
+    "Get weather alerts for a state",
+    {
+      state: z.string().length(2).describe("Two-letter state code (e.g. CA, NY)"),
+    },
+    async ({ state }) => {
+      const stateCode = state.toUpperCase();
+      const alertsUrl = `/alerts?area=${stateCode}`;
+      const alertsData = await makeNWSRequest<AlertsResponse>(alertsUrl);
+
+      if (!alertsData) {
+        return {
+          content: [
+            {
+              type: "text",
+              text: "Failed to retrieve alerts data",
+            },
+          ],
+        };
+      }
+
+      const features = alertsData.features || [];
+      if (features.length === 0) {
+        return {
+          content: [
+            {
+              type: "text",
+              text: `No active alerts for ${stateCode}`,
+            },
+          ],
+        };
+      }
+
+      const formattedAlerts = features.map(formatAlert);
+      const alertsText = `Active alerts for ${stateCode}:\n\n${formattedAlerts.join("\n")}`;
+
+      return {
+        content: [
+          {
+            type: "text",
+            text: alertsText,
+          },
+        ],
+      };
+    },
+  );
+}
